Font layouting and rendering used integers, Xaml layouting and rendering used doubles, everything else used floats - how annoying. Now everything is based on floats. This should also allow smooth font animations, once that is implemented. Some additional cleanup.

+++ Source/Pegasus/Assets Compiler/Fonts/Font.cs
@@ -132,7 +132,7 @@
 
 			var kerning = new FreeType.Vector();
 			FreeType.Invoke(() => FreeType.GetKerning(_fontPtr, left.Index, right.Index, 0, out kerning));
-			return (int)kerning.x / 64;
+			return kerning.x / 64;
 		}
 	}
 }
